deprecate spack solve
